requirements.txt
discord.py>=2.3.0
nudenet>=3.4.2
Pillow>=10.0.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
jinja2>=3.1.0
python-multipart>=0.0.9
httpx>=0.27.0
itsdangerous>=2.1.0
